+++ src/test/kotlin/no/nav/helse/modell/command/nyny/PersonCommandTest.kt
@@ -44,7 +44,7 @@ internal class PersonCommandTest {
         context.add(HentInfotrygdutbetalingerLøsning(objectMapper.createObjectNode()))
         assertTrue(command.execute(context))
         assertFalse(context.harBehov())
-        verify(exactly = 1) { dao.insertPerson(FNR, AKTØR.toLong(), any(), any(), any()) }
+        verify(exactly = 1) { dao.insertPerson(FNR, AKTØR, any(), any(), any()) }
     }
 
     @Test
